requirements.txt
asgiref==3.2.3
autopep8==1.5
Babel==2.8.0
boto3==1.12.39
botocore==1.15.39
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.3
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.5.4
django-choices==1.7.1
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-environ==0.4.5
django-fsm==2.7.0
django-health-check==3.12.1
django-import-export==2.0.2
django-js-asset==1.2.2
django-money==1.0
django-phonenumber-field==4.0.0
django-silk==4.0.0
django-storages==1.9.1
django-tinymce==2.8.0
djangorestframework==3.11.0
docutils==0.15.2
environ==1.0
et-xmlfile==1.0.1
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-core==1.3.0
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gprof2dot==2019.11.30
gunicorn==20.0.4
idna==2.8
jdcal==1.4.1
Jinja2==2.11.1
jmespath==0.9.5
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openpyxl==3.0.3
phonenumbers==8.12.2
Pillow==7.1.1
protobuf==3.11.3
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
Pygments==2.5.2
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.5
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
rsa==4.0
s3transfer==0.3.3
sentry-sdk==0.14.1
six==1.14.0
sqlparse==0.3.0
tablib==1.1.0
urllib3==1.25.8
xlrd==1.2.0
xlwt==1.3.0
